Testes na implementação de mensagens de erro

+++ GestorRestaurante/backend/controllers/UserController.php
@@ -45,16 +45,6 @@ class UserController extends Controller
         ];
     }
 
-    /**
-     * Lists all User models.
-     * @return mixed~
-    */
-    public function actionLogout()
-    {
-        Yii::$app->user->logout();
-
-        return $this->goBack();
-    }
     public function actionIndex()
     {
         $users = User::find()->all();
@@ -68,13 +58,6 @@ $            'dataProvider' => dataProvider,*/
         ]);
     }
 
-    /**
-     * Displays a single User model.
-     *
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionView($id)
     {
 
@@ -90,36 +73,22 @@ $            'dataProvider' => dataProvider,*/
         ]);
     }
 
-    /**
-     * Creates a new User model.
-     * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return mixed
-     */
     public function actionCreate()
     {
         $model = new SignupForm();
         if ($model->load(Yii::$app->request->post()) && $model->signup()) {
             Yii::$app->session->setFlash('success', 'Thank you for registration. Please check your inbox for verification email.');
-            return $this->goHome();
-        }
 
+            return $this->redirect(['index']);
+        }
 
         $model->createAt = date('Y-m-d H:i:s');
         $model->updateAt=date('Y-m-d H:i:s');
-
 
         return $this->render('create', [
             'model' => $model,
         ]);
     }
-
-    /**
-     * Updates an existing User model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
 
     public function actionUpdate($id)
     {
@@ -196,31 +165,30 @@ $            'dataProvider' => dataProvider,*/
             'user' => $user, 'perfil' =>$perfil
         ]);
     }
-    /**
-     * Deletes an existing User model.
-     * If deletion is successful, the browser will be redirected to the 'index' page.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
+
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (Yii::$app->user->can('apagarUtilizadores')) {
 
-        return $this->redirect(['index']);
+            if(Yii::$app->authManager->checkAccess())
+                Perfil::findOne($id)->delete();
+
+            $this->findModel($id)->delete();
+
+            return $this->redirect(['index']);
+
+        }else{
+
+            return $this->render('site/error');
+        }
+
+
     }
 
-    /**
-     * Finds the User model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param integer $id
-     * @return User the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     protected function findModel($id)
     {
-        if (($model = User::findOne($id)) !== null) {
-            return $model;
+        if (($user = User::findOne($id)) !== null) {
+            return $user;
         }
 
         throw new NotFoundHttpException('The requested page does not exist.');
@@ -234,7 +202,6 @@ $            'dataProvider' => dataProvider,*/
 
 
     }
-
 
     public function actionRemovecargo($cargo,$id_user){
